+++ ofrak_core/src/ofrak/gui/server.py
@@ -1108,6 +1108,9 @@ class AiohttpOFRAKServer:
                         else None,
                     }
                 )
+            docstring = "No documentation is provided for this config."
+            if config.__doc__ is not None:
+                docstring = config.__doc__
             return json_response(
                 {
                     "name": config.__name__,
@@ -1116,6 +1119,7 @@ class AiohttpOFRAKServer:
                     "args": self._construct_arg_response(self._convert_to_class_name_str(config)),
                     "enum": self._construct_enum_response(config),
                     "fields": _fields,
+                    "doc": docstring,
                 }
             )
         else: